fix: eliminar JAVA-W1052 reemplazando intValue() con Long en repositorio
- fix: cambiar repository methods a Long (W1052 unboxing)
- fix: actualizar adapter y tests para usar Long
- fix: completar ProjectControllerTest con mock de rol

+++ fiis/src/main/java/com/sgi/fiis/tramites/infrastructure/persistence/ProcedureMovementRepositoryAdapter.java
@@ -16,7 +16,7 @@ public class ProcedureMovementRepositoryAdapter implements ProcedureMovementRepo
 
     @Override
     public List<ProcedureMovement> findByProcedureId(Long procedureId) {
-        return repository.findByProcedure_IdOrderByMovementAtAsc(procedureId.intValue())
+        return repository.findByProcedure_IdOrderByMovementAtAsc(procedureId)
                 .stream()
                 .map(this::toDomain)
                 .toList();
